pyproject.toml
[tool.poetry]
name = "test-task"
version = "0.1.0"
description = "Testwork for access to the 'Career center'"
authors = ["Aleksey Lesnikov"]
readme = "README.md"
packages = [{include = "test_task"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
envparse = "^0.2.0"
psycopg2-binary = "^2.9.5"
djangorestframework = "^3.14.0"
django-filter = "^22.1"
drf-spectacular = "^0.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
